fix compiling errors due to the latest xamarin.iOS framework update

+++ SlidingPanels/SlidingPanels.Lib/Tools/UIImageEffects.cs
@@ -89,20 +89,20 @@ namespace UIImageEffects
 						using (var effectInContext = contextIn.AsBitmapContext () as CGBitmapContext) {
 							var effectInBuffer = new vImageBuffer () {
 								Data = effectInContext.Data,
-								Width = effectInContext.Width,
-								Height = effectInContext.Height,
-								BytesPerRow = effectInContext.BytesPerRow
+								Width = (int)effectInContext.Width,
+								Height = (int)effectInContext.Height,
+								BytesPerRow = (int)effectInContext.BytesPerRow
 							};
-							
+
 							UIGraphics.BeginImageContextWithOptions (image.Size, false, currentScale);
 							
 							try {
 								using (var effectOutContext = UIGraphics.GetCurrentContext ().AsBitmapContext () as CGBitmapContext) {
 									var effectOutBuffer = new vImageBuffer () {
 										Data = effectOutContext.Data,
-										Width = effectOutContext.Width,
-										Height = effectOutContext.Height,
-										BytesPerRow = effectOutContext.BytesPerRow
+										Width = (int)effectOutContext.Width,
+										Height = (int)effectOutContext.Height,
+										BytesPerRow = (int)effectOutContext.BytesPerRow
 									};
 
 									if (hasBlur) {
